+++ packages/atom-react-redux/src/index.ts
@@ -22,6 +22,7 @@ export type { ValidatableCallbackOptions } from "./withValidatableCallback";
 // callApi saga builder
 export { callApi } from "./callApi";
 export { ApiCallResult } from "./callApi";
+export type { VersionConflictHandling } from "./callApi";
 
 // Common slice (for use in FakeStore integration)
 export * as Common from "./common";